### 2022-10-21 CliMA/EnsembleKalmanProcesses.jl#222 @odow
JOSS paper review
Hi all, I'm reviewing the JOSS paper: https://github.com/openjournals/joss-reviews/issues/4869#issuecomment-1286002558

A couple of things jumped out at me, going off the JOSS checklist:

 * **A statement of need** The docs start with
   > EnsembleKalmanProcesses.jl (EKP) is a library of derivative-free Bayesian optimization techniques based on Ensemble Kalman Filters, a well known family of approximate filters used for data assimilation. Currently, the following methods are implemented in the library:

   and then it goes into academic literature review. I think you could start with a more generic statement of need that highlighted the derivative free optimization and its applicability for fitting parameters to expensive black-box simulation data.

 * **Example usage** The doc order could change. Currently, the examples are right at the bottom, and things like the contributing guide are higher. Provide a simple example as early as possible. Compare the BayesianOptimization intro examples: https://github.com/fmfn/BayesianOptimization#basic-tour-of-the-bayesian-optimization-package

* **State of the field** you could briefly position EKP in the context of other Bayesian and DFO tools like BlackBoxOptim.jl and Python tools like BayesianOptimization (and presumably other EKP-like packages?)

The paper is also quite long. I'd be in favor of shrinking the "features" section in favor of a small code example that demonstrated the package.

Thoughts?
