### 2022-10-21 CliMA/EnsembleKalmanProcesses.jl#222 @odunbar
Hi @odow!

Thanks for the rapid feedback, I agree with your comments, particularly that an immediate visual could help users understand the package,

Here's a first list to address your comments.
_In the Docs_:
-  A statement of need can be added to the landing page
-  A visually attractive simulation of a pedagogical example can be added to the landing page
(e.g. learning some amplitude/shifts of a randomly perturbed sine wave, from a simple measurement)
- Re-ordering of the contents to promote examples

_In the paper_:
- A statement to position our algorithms with regards to `BlackBoxOptim.jl` and other suitable packages we can find.
- Add code snippets/figures from an example (e.g. the pedagogical example mentioned above)

Regarding the length, I've seen JOSS papers range from 1 paragraph to 6 pages, so I'm not sure if i necessarily agree that the length is extreme, but I know cutting can improve focus and conciseness. When you mention "shrinking", would you prefer less verbose language, or do you have specific pieces you feel readers would not miss being cut.
