### 2022-10-23 CliMA/EnsembleKalmanProcesses.jl#222 @odow
I'd also move the statement of need to after the summary and before the "Features" section.

> i necessarily agree that the length is extreme

I don't know if I said extreme, but it's heading towards the upper end, especially if you add an example with a figure.

> When you mention "shrinking", would you prefer less verbose language, or do you have specific pieces you feel readers would not miss being cut.

I thought lines 46-61 could be cut in favor of a sentence like "We also implement a number of features recently described in the literature to improve the robustness and flexibility of the ensemble algorithms (cites...)." But perhaps add the example etc, the see how the length is before cutting.
