### 2022-10-24 CliMA/EnsembleKalmanProcesses.jl#222 @odunbar
Thanks for the update and clarification. We will start working on this soon.
