### 2022-11-03 9ebf66b
Adding sinusoid example
