### 2022-11-10 c1949a2
add detail to introduction
draft comparison to other tools

paper.md

ilopezgp comments

texttt on EKP

added kalman method packages

refined statement, added ref

typo

small edit

typo

Update paper.md

Co-authored-by: Ignacio Lopez-Gomez <ilopezgo@caltech.edu>

add pdf
