+++ paper.bib
@@ -218,3 +218,18 @@ author = {Emmet Cleary and Alfredo Garbuno-Inigo and Shiwei Lan and Tapio Schnei
     year     = {2022},
     url = {https://hdl.handle.net/1721.1/145140}
 }
+
+
+
+@article{Schillings2017,
+author = {Schillings, Claudia and Stuart, Andrew M.},
+title = {Analysis of the Ensemble Kalman Filter for Inverse Problems},
+journal = {SIAM Journal on Numerical Analysis},
+volume = {55},
+number = {3},
+pages = {1264-1290},
+year = {2017},
+doi = {10.1137/16M105959X},
+}
+
+
