+++ paper.md
@@ -41,11 +41,41 @@ bibliography: paper.bib
 
 # Summary
 
-EnsembleKalmanProcesses.jl is a Julia-based toolbox that can be used for a broad class of black-box gradient-free optimization problems. Specifically, the tools enable the optimization, or calibration, of parameters within a computer model in order to best match user-defined outputs of the model with available observed data [@kennedy_ohagan_2001]. Some of the tools can also approximately quantify parametric uncertainty [@Huang:2022b]. Though the package is written in Julia, a read–write TOML-file interface is provided so that the tools can be applied to computer models implemented in any language. Furthermore, the calibration tools are non-intrusive, relying only on the ability of users to compute an output of their model given a parameter value.
+\texttt{EnsembleKalmanProcesses.jl} is a Julia-based toolbox that can be used for a broad class of black-box gradient-free optimization problems. Specifically, the tools enable the optimization, or calibration, of parameters within a computer model in order to best match user-defined outputs of the model with available observed data [@kennedy_ohagan_2001]. Some of the tools can also approximately quantify parametric uncertainty [@Huang:2022b]. Though the package is written in Julia, a read–write TOML-file interface is provided so that the tools can be applied to computer models implemented in any language. Furthermore, the calibration tools are non-intrusive, relying only on the ability of users to compute an output of their model given a parameter value.
 
 As the package name suggests, the tools are inspired by the well-established class of ensemble Kalman methods. Ensemble Kalman filters are currently one of the only practical ways to assimilate large volumes of observational data into models for operational weather forecasting [@Evensen:1994;@Houtekamer:1998;@Houtekamer:2001]. In the data assimilation setting, a computational weather model is integrated for a short time over a collection, or ensemble, of initial conditions, and the ensemble is updated frequently by a variety of atmospheric observations, allowing the forecasts to keep track of the real system.
 
-The workflow is similar for ensemble Kalman processes. Here, a computer code is run (in parallel) for an ensemble of different values of the parameters that require calibration, producing an ensemble of outputs. This ensemble of outputs is then compared to observed data, and the parameters are updated to a new set of values which reduce the output–data misfit. The computer model is then evaluated for the new ensemble values and the outputs, under certain conditions, are guaranteed to lie closer to the observed data. The process is iterated until a user-defined criterion of convergence is met. Optimal values are selected from statistics of the final ensemble.
+The workflow is similar for ensemble Kalman processes. Here, a computer code is run (in parallel) for an ensemble of different values of the parameters that require calibration, producing an ensemble of outputs. This ensemble of outputs is then compared to observed data, and the parameters are updated to a new set of values which reduce the output–data misfit. The computer model is then evaluated for the new ensemble values and the outputs. Optimality of the update is guaranteed for linear models and Gaussian uncertainties, but good performance is observed outside of these settings, see [@Schillings2017]. The process is iterated until a user-defined criterion of convergence is met. Optimal values are selected from statistics of the final ensemble.
+
+# Statement of need
+
+The task of estimating parameters of a computer model or simulator such that its outputs fit with data is ubiquitous in science and engineering, coming under many names such as calibration, inverse problems, and parameter estimation. In statistics and machine learning, when closed-form estimators of parameters of a model are unavailable, similar approaches may need to be employed to fit the model to data. There is a wide variety of algorithms to suit these applications; however, there are many bottlenecks in the practical application of such methods to computer codes:
+
[…]
