### 2022-11-14 CliMA/EnsembleKalmanProcesses.jl#232 @ziyiyin97
JOSS Paper Review 2
My apologies for the late reviews https://github.com/openjournals/joss-reviews/issues/4869

Overall the package shows how to run derivative-free ensemble kalman process in Julia and entails quite clean documentation and examples. There are also a lot of useful references in the manuscript. A few comments:

1. In general how does this package compare to other Kalman-related packages? like Kalman.jl or other Julia/Python based packages?

Comments for the manuscript:
1. Line 30 "under certain conditions" -- probably need a reference.

In terms of the code/examples/docs:
1. The package claims that the prior distribution can be either from Julia Distributions or empirical/sample-based distributions. I would appreciate it if authors could provide an example related to sample-based distributions. (or point to reference)
2. Also appreciate an example w/ TOML I/O for a few iterations. (or point to reference) I assume these are both related to this module https://github.com/CliMA/EnsembleKalmanProcesses.jl/blob/main/src/UQParameters.jl

I hope these concerns can be easily addressed with more explanations or so.
