### 2022-11-15 CliMA/EnsembleKalmanProcesses.jl#232 @odunbar
Hi @ziyiyin97 Thank you very much for your review.

Proposed response:
_In paper.md_
- We are currently writing a state-of-field section for the other review, and where our code fits in, we shall include `Kalman.jl` in this.
- "under certain conditions" refers to the standard Kalman assumptions, we shall tighten up this statement.

_In the docs_
- We will add a docs for defining a sample-based distribution.
- We will add a docs tutorial for the TOML interface

_In the code_
- We will add a simple example that uses the TOML interface

We shall keep you updated! Thanks again
