### 2022-11-18 CliMA/EnsembleKalmanProcesses.jl#222 @odunbar
Hi @odow Thank you for your patience! Our changes:
- PR #228 adds a pedagogical example to the documentation, and also displaying a gif of its results on the docs home-page.
- PR #231 adds a state of field to paper.md and rearranges the sections, adds more in depth statement to docs home-page
- PR #234 reorders documentation to promote examples
Left to do:
- Add code-snippets/example into paper.md
