### 2022-11-18 96791de
add first draft sinusoid in paper
code -> tex

format

format

format

julia syntax

julia syntax

caption

remove 1,7, change \phi

rewording

capitalization

review

corrected final paragraph

paper.md

added solution explanation

typo and reword

added pdf

last typo...

add pdf
