### 2022-11-21 f9df438
add particle inflation functionality
