### 2022-11-22 CliMA/EnsembleKalmanProcesses.jl#222 @odow
Looks good. I wonder if it's still a little long, but I'll leave that to the editors if they complain.
