### 2022-11-23 478ef82
rename UQParameters to TOMLInterface
rename

new example

format

readme

readme

unit tests

consistent with interface

unit tests pass

compatible examples

format

add end message

codecov

rename UQ -> admissibe

format

added API page

change scope

example docs

format

link to parallel page

moved files from EKP to reduce compile time a bit

added priors.toml into docs

remove old file

Update Docs from code review

Co-authored-by: Ignacio Lopez-Gomez <ilopezgo@caltech.edu>
Co-authored-by: Eviatar Bach <eviatarbach@protonmail.com>

rm plots completely

docs snippet change, & 6 ens membe
[…]
