+++ examples/SinusoidInterface/shared.jl
@@ -1,0 +1,8 @@
+# shared information
+
+using LinearAlgebra, Random, TOML, JLD2
+using Distributions
+using EnsembleKalmanProcesses
+using EnsembleKalmanProcesses.TOMLInterface
+using EnsembleKalmanProcesses.ParameterDistributions
+const EKP = EnsembleKalmanProcesses
