+++ docs/src/API/TOMLInterface.md
@@ -1,0 +1,15 @@
+# TOML interface
+
+```@meta
+CurrentModule = EnsembleKalmanProcesses.TOMLInterface
+```
+
+```@docs
+path_to_ensemble_member
+get_parameter_distribution
+get_parameter_values
+save_parameter_ensemble
+get_admissible_parameters
+get_regularization
+write_log_file
+```
