+++ examples/FMS/helper_funcs.jl
@@ -3,6 +3,7 @@ using Statistics
 using Interpolations
 using LinearAlgebra
 using JLD
+using EnsembleKalmanProcesses
 
 """
 Transfer function 
@@ -11,10 +12,39 @@ Transfer function
 
 
 params_prefix =  "output/output_params_"
-input_prefix  =  "output/output"
-output_prefix =  "output/output"
-N_params = 2
+input_prefix  =  "output/output_"
+output_prefix =  "output/output_"
+input_file_template = "linear.jl"
+
+u_name = ["theta", "theta"]
+
+prior_mean = [0.0; 0.0]
+prior_cov  = [1.0 0.0; 0.0 1.0]
+
+
+obs_mean, obs_noise_cov = [3], 0.1^2*[1.0]
+
+G = [ 1 2]
+u = [ 1; 1]
+y = G*u
+uu_cov_post = inv(G'*(obs_noise_cov\G) + inv(prior_cov))
+
+u_refs = u_mean_post = prior_mean + uu_cov_post*(G'*(obs_noise_cov\(y - G*prior_mean)))
+
+
+
+
+
+N_params, N_y = 2, 1+2
 param_bounds = [[nothing, nothing], [nothing, nothing]]
+
+
+
+
+
+α_reg = 1.0 
+update_freq = 1 
+sigma_points_type = "symmetric"
 
 N_ens = 2N_params  + 1
 
@@ -31,6 +61,17 @@ function constraint(u::Float64, u_low, u_up)
     end
 end
 
+function dconstraint(u::Float64, u_low, u_up)
+    if(isnothing(u_low) && isnothing(u_up))
+        return 1
+    elseif (isnothing(u_low)  &&  !isnothing(u_up))
+    	return -exp(u)
+    elseif (!isnothing(u_low)  &&  isnothing(u_up))
+        return exp(u)
+    else
+        return  -(u_up - u_low)*exp(u)/(1 + exp(u))^2
+    end
+end
 
 function constraint(u::Array{Float64, 1})
   constraint_u = similar(u)
@@ -40,13 +81,23 @@ function constraint(u::Array{Float64, 1})
   return constraint_u
 end
 
+function constraint(u_mean::Array{Float64, 1}, uu_cov::Array{Float64, 2})
+  N_params = length(u_mean)
+  dc = zeros(N_params, N_params)
+  for i = 1:N_params
+      dc[i,i] = dconstraint(u_mean[i], param_bounds[i][1], param_bounds[i][2])
+  end
+  
+  constraint_uu_cov = dc * uu_cov * dc'
+  return constraint_uu_cov
+end
 
 function constraint(u_ens::Array{Float64, 2})
   constraint_u_ens = similar(u_ens)
   N_params, N_ens = size(u_ens)
 
-  for i = 1:N
-      constraint_u_ens[:, i] = constraint_helper(u_ens[:, i])
+  for i = 1:N_ens
+      constraint_u_ens[:, i] = constraint(u_ens[:, i])
   end
 
   return constraint_u_ens
@@ -56,20 +107,44 @@ end
 # save mean and covariance and u_ens
 function save_params(ukiobj, iteration_::Int64)
     mean, cov = ukiobj.process.u_mean[end], ukiobj.process.uu_cov[end]
-    u_ens = construct_sigma_ensemble(ukiobj.process, mean, cov)
-    u_p_ens = get_u_final(ukobj)
+    u_ens = EnsembleKalmanProcesses.construct_sigma_ensemble(ukiobj.process, mean, cov)
+    u_p_ens = get_u_final(ukiobj)
          
[…]
