+++ examples/FMS/preprocess.jl
@@ -10,8 +10,9 @@ include(joinpath(@__DIR__, "helper_funcs.jl"))
 
 
 run(`mkdir -p output`)
-run(`cd output`)
+run(`mkdir -p output/slurm`)
 for i = 1:N_ens
-    run(`mkdir -p output_fms_$(i)`)
+    run(`mkdir -p $(input_prefix)$(i)`)
+    run(`mkdir -p $(output_prefix)$(i)`)
 end
 
