+++ examples/FMS/helper_funcs.jl
@@ -9,43 +9,13 @@ using EnsembleKalmanProcesses
 Transfer function 
 """
 
-
-
-params_prefix =  "output/output_params_"
-input_prefix  =  "output/output_"
-output_prefix =  "output/output_"
-input_file_template = "linear.jl"
-
-u_name = ["theta", "theta"]
-
-prior_mean = [0.0; 0.0]
-prior_cov  = [1.0 0.0; 0.0 1.0]
-
-
-obs_mean, obs_noise_cov = [3], 0.1^2*[1.0]
-
-G = [ 1 2]
-u = [ 1; 1]
-y = G*u
-uu_cov_post = inv(G'*(obs_noise_cov\G) + inv(prior_cov))
-
-u_refs = u_mean_post = prior_mean + uu_cov_post*(G'*(obs_noise_cov\(y - G*prior_mean)))
-
-
-
-
-
-N_params, N_y = 2, 1+2
-param_bounds = [[nothing, nothing], [nothing, nothing]]
-
-
+include("helper_funcs_linear.jl")
 
 
 
 α_reg = 1.0 
 update_freq = 1 
 sigma_points_type = "symmetric"
-
 N_ens = 2N_params  + 1
 
 
@@ -139,60 +109,5 @@ function read_all_params(N_iterations::Int64)
     return constraint_u_mean_all, constraint_uu_cov_all
 end
 
-function read_observation()
-
-    obs_mean, obs_noise_cov = [3], 0.1^2*[1.0]
-    N_y = length(obs_mean)
-    return [obs_mean; prior_mean], [obs_noise_cov zeros(Float64, N_y, N_params); zeros(Float64, N_params, N_y) prior_cov]
-end
-
-function write_fms_input_files(constraint_u_p_ens)
-
-    N_params, N_ens = size(constraint_u_p_ens)
-
-    input_lines = readlines(input_file_template)
-    
-    modified_lines = [5; 8]
-    for i = 1:N_ens
-    	constraint_u_p = constraint_u_p_ens[:, i]
- 	input_name = input_prefix*string(i)*"/"*"input_file"
-        output_name = output_prefix*string(i)*"/"*"output_file.jld"
- 
-    	input_io = open(input_name, "w")
-	 	
-    	for (n, line) in enumerate(input_lines)
-       	   
-	   if n ∉ modified_lines
-
-              write(input_io, line*"\n")
-
-           else
-	      if n == 5
-	      	 write(input_io, "u = [$(constraint_u_p[1]) ;  $(constraint_u_p[2])] \n")
-	      elseif n == 8
-	      	 write(input_io, "save(\"$(output_name)\", \"y\" , y) \n")    
-	      else
-		  @error("STOP: line $(n) should not be modified!")
-	      end              
-	   
-	   end
-
-       end
-       
-       close(input_io)
-
-    end
-
-
-end
-
-
-function read_fms_output(iteration_::Int64,  ens_index::Int64)
-    _, _, _, u_p_ens = read_params( iteration_)
-    
-    data = load(output_prefix*string(ens_index)*"/"*"output_file.jld")
-    
-    return [data["y"]; u_p_ens[:, ens_index]]
-end
 
 
