### 2022-11-28 CliMA/EnsembleKalmanProcesses.jl#232 @odunbar
An update of where we are:
- PR #231 adds a state-of-field, and tightened Kalman assumptions (along with changes for the other review)
- PR #238 adds a replica of our simple sinusoid example, but uses the TOML and file-system interface between model and calibration tools. [Under Code Review]
- PR #240 adds some simple intuitive examples for creating prior distributions using "samples-based" distributions into the docs. [Under Code Review]
I'll ping you again when these are complete
