### 2022-11-28 #4869 @odunbar
An update:

The 3 PRs in [this comment](https://github.com/CliMA/EnsembleKalmanProcesses.jl/issues/232#issuecomment-1329881988) are addressing the second reviewers comments.

The PRs are undergoing the final internal review, then I shall ask the second reviewer to take a look.
