### 2022-11-29 CliMA/EnsembleKalmanProcesses.jl#232 @odunbar
@ziyiyin97 I believe we have now addressed all the issues raised.
- The changes to the paper are in [paper.pdf](https://github.com/CliMA/EnsembleKalmanProcesses.jl/files/10117338/paper.pdf)
- The new TOML example is found in [`examples/SinusoidInterface/`](https://github.com/CliMA/EnsembleKalmanProcesses.jl/tree/main/examples/SinusoidInterface)
- The docs for the TOML examples [here](https://clima.github.io/EnsembleKalmanProcesses.jl/dev/examples/sinusoid_example_toml/#TOML-interface-for-fitting-parameters-of-a-sinusoid)
- The docs for the samples-example [here](https://clima.github.io/EnsembleKalmanProcesses.jl/dev/parameter_distributions/#samples-example)
