### 2022-11-30 CliMA/EnsembleKalmanProcesses.jl#232 @ziyiyin97
Thanks for your quick update. The new TOML example in https://github.com/CliMA/EnsembleKalmanProcesses.jl/tree/main/examples/SinusoidInterface needs an updated version of EnsembleKalmanProcess otherwise `julia --project` will use the previous version of EnsembleKalmanProcess w/o `TOMLInterface` --i.e. `ERROR: UndefVarError: TOMLInterface not defined`. But everything else looks good to me and this issue can be fixed by burning a new release. Thanks again for your great work which enriches the Kalman community in Julia!
