### 2022-12-04 #4869 @jbytecode
Dear @odunbar

I've just sent a pull request: https://github.com/CliMA/EnsembleKalmanProcesses.jl/pull/242

Please review the changes and apply the PR if you agree with them.

Thank you in advance.
