### 2022-12-05 4e7e6ec
Some typo fixes, and paragraph shortening.
