+++ paper.bib
@@ -37,6 +37,20 @@
   journal   = {{SIAM} Review}
 }
 
+@inproceedings{ge2018t,
+  author    = {Hong Ge and
+               Kai Xu and
+               Zoubin Ghahramani},
+  title     = {Turing: a language for flexible probabilistic inference},
+  booktitle = {International Conference on Artificial Intelligence and Statistics,
+               {AISTATS} 2018, 9-11 April 2018, Playa Blanca, Lanzarote, Canary Islands,
+               Spain},
+  pages     = {1682--1690},
+  year      = {2018},
+  url       = {http://proceedings.mlr.press/v84/ge18b.html},
+  biburl    = {https://dblp.org/rec/bib/conf/aistats/GeXG18},
+}
+
 @article{Lopez-Gomez:2022,
 author = {Lopez-Gomez, Ignacio and Christopoulos, Costa and Langeland Ervik, Haakon Ludvig and Dunbar, Oliver R. A. and Cohen, Yair and Schneider, Tapio},
 title = {Training Physics-Based Machine-Learning Parameterizations With Gradient-Free Ensemble {Kalman} Methods},
