+++ paper.bib
@@ -6,7 +6,7 @@
   number = {78}, 
   pages = {4723}, 
   author = {Jesús-Adolfo Mejía-de-Dios and Efrén Mezura-Montes}, 
-  title = {Metaheuristics: A Julia Package for Single- and Multi-Objective Optimization}, 
+  title = {Metaheuristics: A {Julia} Package for Single- and Multi-Objective Optimization}, 
   journal = {Journal of Open Source Software} 
 }
 
@@ -19,7 +19,7 @@
   number = {24}, 
   pages = {615}, 
   author = {Patrick K. Mogensen and Asbjørn N. Riseth}, 
-  title = {Optim: A mathematical optimization package for Julia}, 
+  title = {Optim: A mathematical optimization package for {Julia}}, 
   journal = {Journal of Open Source Software} 
 }
 
