### 2022-12-05 #4869 @jbytecode
@odunbar  - The PR (https://github.com/CliMA/EnsembleKalmanProcesses.jl/pull/244) includes minor changes in the bibtex files. Please apply them. Thank you in advance.
