### 2022-12-06 #4869 @odunbar
I've completed the checklist,
- I've read through and I am happy with it.
- Tagged release:
[v0.14.1](https://github.com/CliMA/EnsembleKalmanProcesses.jl/releases/tag/v0.14.1)
- Zenodo archive of tagged release with consistent metadata:
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7407193.svg)](https://doi.org/10.5281/zenodo.7407193)
