### 2022-12-07 #4869 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
