### 2022-12-14 #4869 @odunbar
Hi @jbytecode, do you know how long this stage typically takes?
